refactor code to move anuth code to a separate component
