+++ src/App.js
@@ -1,14 +1,11 @@
 import React, { useState, useEffect } from 'react';
 import './App.css';
+import Auth from './components/Auth';
 
 const DEFAULT_PIN = '0000';
-const MAX_ATTEMPTS = 3;
 
 function App() {
   const [authenticated, setAuthenticated] = useState(false);
-  const [pin, setPin] = useState('');
-  const [pinError, setPinError] = useState('');
-  const [attempts, setAttempts] = useState(0);
 
   const [items, setItems] = useState(() => {
     const saved = localStorage.getItem('testElements');
@@ -22,10 +19,6 @@ function App() {
   const [storedPin, setStoredPin] = useState(() => {
     return localStorage.getItem('appPin') || DEFAULT_PIN;
   });
-  const [showChangePin, setShowChangePin] = useState(false);
-  const [oldPin, setOldPin] = useState('');
-  const [newPin, setNewPin] = useState('');
-  const [confirmPin, setConfirmPin] = useState('');
   const [changePinMsg, setChangePinMsg] = useState('');
 
   // Delete confirmation
@@ -95,128 +88,15 @@ function App() {
     }
   };
 
-  const handlePinSubmit = () => {
-    if (pin === storedPin) {
-      setAuthenticated(true);
-      setPinError('');
-      setPin('');
-      setAttempts(0);
-    } else {
-      const newAttempts = attempts + 1;
-      setAttempts(newAttempts);
-      if (newAttempts >= MAX_ATTEMPTS) {
-        setPinError('You have reached the maximum number of attempts. Access denied.');
-      } else if (newAttempts === MAX_ATTEMPTS - 1) {
-        setPinError('This is your last attempt!');
-      } else {
-        setPinError('Incorrect PIN. Try again.');
-      }
-    }
-    setPin('');
-  };
-
-  // Change password logic
-  const handleChangePin = () => {
-    setChangePinMsg('');
-    if (oldPin !== storedPin) {
-      setChangePinMsg('Old PIN is incorrect.');
-      return;
-    }
-    if (newPin.length !== 4 || isNaN(newPin)) {
-      setChangePinMsg('New PIN must be a 4-digit number.');
-      return;
-    }
-    if (newPin !== confirmPin) {
-      setChangePinMsg('New PIN and confirm PIN do not match.');
-      return;
-    }
-    setStoredPin(newPin);
-    setShowChangePin(false);
-    setOldPin('');
-    setNewPin('');
-    setConfirmPin('');
-    setChangePinMsg('PIN changed successfully!');
-    setTimeout(() => setChangePinMsg(''), 2000);
-  };
-
   if (!authenticated) {
-    const isLocked = attempts >= MAX_ATTEMPTS;
     return (
-      <div className="App">
-        <h2>Enter PIN to Access</h2>
-        <input
-          type="password"
-          value={pin}
-          onChange={e => setPin(e.target.value)}
-          placeholder="Enter PIN"
-          maxLength={4}
-          onKeyDown={e => {
-            if (e.key === 'Enter' && !isLocked) handlePinSubmit();
-          }}
-          disabled={isLocked}
-        />
-        <button onClick={handlePinSubmit} disabled={isLocked}>Submit</button>
-        {pinError && <p style={{ color: 'red' }}>{pinError}</p>}
-        <div style={{ width: '100%', display: 'flex', justifyContent: 'center', alignItems: 'center', marginTop: 12 }}>
-          <button
-            style={{
-              background: 'transparent',
-              border: 'none',
-              color: '#61dafb',
-              fontSize: '1em',
-              cursor: 'pointer',
-              textDecoration: 'underline'
-            }}
-            title="Change PIN"
-            onClick={() => setShowChangePin(!showChangePin)}
-          >
-            Change PIN
-          </button>
-        </div>
-        {showChangePin && (
-          <div
-            className="change-pin"
-            style={{
-              background: '#23262f',
-              borderRadius: 8,
-              padding: 16,
-              marginTop: 16,
-              minWidth: 260
-            }}
-          >
-            <h3 style={{ marginTop: 0 }}>Change PIN</h3>
-            <input
-              type="password"
-              value={oldPin}
-              onChange={e => setOldPin(e.target.value)}
-              placeholder="Old PIN"
-              maxLength={4}
-              style={{ marginBottom: 8 }}
-            />
-            <input
-              type="password"
-              value={newPin}
-              onChange={e => setNewPin(e.target.value)}
-              placeholder="New PIN"
-              maxLength={4}
-              style={{ marginBottom: 8 }}
-            />
-            <input
-              type="password"
-              value={confirmPin}
-              onChange={e => setConfirmPin(e.target.value)}
-              placeholder="Confirm New PIN"
-              maxLength={4}
-              style={{ marginBottom: 8 }}
-            />
-            <div>
-              <button onClick={handleChangePin}>Save</button>
-              <button onClick={() => setShowChangePin(false)} style={{ marginLeft: 8 }}>Cancel</button>
-            </div>
-            {changePinMsg && <p style={{ color: changePinMsg.includes('success') ? 'lightgreen' : 'red' }}>{changePinMsg}</p>}
-          </div>
-        )}
-      </div>
+      <Auth
+        storedPin={storedPin}
+        setStoredPin={setStoredPin}
+        setAuthenticated={setAuthenticated}
+        changePinMsg={changePinMsg}
+        setChangePinMsg={setChangePinMsg}
+      />
     );
   }
 
@@ -236,94 +116,91 @@ function App() {
         />
         <button onClick={handleAdd}>Add</button>
       </div>
-      {!showChangePin && (
-        <ul>
-          {items.length === 0 && <li>No elements yet.</li>}
-          {[...items].sort((a, b) => a.localeCompare(b)).map((item, idx) => {
-            const origIndex = items.findIndex(i => i === item);
-            return (
-              <li
-                key={idx}
-                style={{
-                  display: 'flex',
-                  alignItems: 'center',
-                  justifyContent: 'space-between',
-                  width: '100%',
-                  position: 'relative'
-                }}
-                className="list-item"
-              >
-                {editIndex === origIndex ? (
-                  <>
-                    <input
-                      type="text"
-                      value={editValue}
-                      onChange={handleEditChange}
-                      onKeyDown={e => handleEditKeyDown(e, origIndex)}
-                      className="edit-input"
-                      autoFocus
-                      style={{ flex: 1, marginRight: 8 }}
-                    />
-                    <button
-                      onClick={() => handleEditSave(origIndex)}
-                      className="save-btn"
-                      title="Save"
-                    >✔️</button>
-                    <button
-                      onClick={() => {
-                        setEditIndex(null);
-                        setEditValue('');
-                      }}
-                      className="cancel-btn"
-                      title="Cancel"
-                    >✖️</button>
-                  </>
-                ) : (
-                  <>
-                    <span style={{ flex: 1 }}>{item}</span>
-                    <button
-                      onClick={() => handleEdit(origIndex)}
-                      className="edit-btn"
-                      title="Edit"
-                      disabled={editIndex !== null}
-                    >✏️</button>
-                    <button
-                      onClick={() => handleRemove(origIndex)}
-                      className="remove-btn"
-                      title="Remove"
-                      disabled={editIndex !== null}
-                    >&#10006;</button>
-                  </>
-                )}
-                {/* Delete confirmation dialog */}
-                {deleteIndex === origIndex && (
-                  <div
-                    style={{
-                      position: 'absolute',
-                      top: '50%',
-                      left: '50%',
-                      transform: 'translate(-50%, -50%)',
-                      background: '#23262f',
-                      color: '#fff',
-                      border: '1px solid #353945',
-                      borderRadius: 8,
-                      padding: 16,
-                      zIndex: 10,
-                      minWidth: 200,
-                      boxShadow: '0 2px 8px rgba(0,0,0,0.3)'
+      <ul>
+        {items.length === 0 && <li>No elements yet.</li>}
+        {[...items].sort((a, b) => a.localeCompare(b)).map((item, idx) => {
+          const origIndex = items.findIndex(i => i === item);
+          return (
+            <li
+              key={idx}
+              style={{
+                display: 'flex',
+                alignItems: 'center',
+                justifyContent: 'space-between',
+                width: '100%',
+                position: 'relative'
+              }}
+              className="list-item"
+            >
+              {editIndex === origIndex ? (
+                <>
+                  <input
+                    type="text"
+                    value={editValue}
+                    onChange={handleEditChange}
+                    onKeyDown={e => handleEditKeyDown(e, origIndex)}
+                    className="edit-input"
+                    autoFocus
+                    style={{ flex: 1, marginRight: 8 }}
+                  />
+                  <button
+                    onClick={() => handleEditSave(origIndex)}
+                    className="save-btn"
+                    title="Save"
+                  >✔️</button>
+                  <button
+                    onClick={() => {
+                      setEditIndex(null);
+                      setEditValue('');
                     }}
-                  >
-                    <div style={{ marginBottom: 12 }}>Are you sure you want to delete this item?</div>
-                    <button onClick={confirmRemove} style={{ marginRight: 8 }}>Yes</button>
-                    <button onClick={cancelRemove}>No</button>
-                  </div>
-                )}
-              </li>
-            );
-          })}
-        </ul>
-      )}
-      {showChangePin && null}
+                    className="cancel-btn"
+                    title="Cancel"
+                  >✖️</button>
+                </>
+              ) : (
+                <>
+                  <span style={{ flex: 1 }}>{item}</span>
+                  <button
+                    onClick={() => handleEdit(origIndex)}
+                    className="edit-btn"
+                    title="Edit"
+                    disabled={editIndex !== null}
+                  >✏️</button>
+                  <button
+                    onClick={() => handleRemove(origIndex)}
+                    className="remove-btn"
+                    title="Remove"
+                    disabled={editIndex !== null}
+                  >&#10006;</button>
+                </>
+              )}
+              {/* Delete confirmation dialog */}
+              {deleteIndex === origIndex && (
+                <div
+                  style={{
+                    position: 'absolute',
+                    top: '50%',
+                    left: '50%',
+                    transform: 'translate(-50%, -50%)',
+                    background: '#23262f',
+                    color: '#fff',
+                    border: '1px solid #353945',
+                    borderRadius: 8,
+                    padding: 16,
+                    zIndex: 10,
+                    minWidth: 200,
+                    boxShadow: '0 2px 8px rgba(0,0,0,0.3)'
+                  }}
+                >
+                  <div style={{ marginBottom: 12 }}>Are you sure you want to delete this item?</div>
+                  <button onClick={confirmRemove} style={{ marginRight: 8 }}>Yes</button>
+                  <button onClick={cancelRemove}>No</button>
+                </div>
+              )}
+            </li>
+          );
+        })}
+      </ul>
     </div>
   );
 }
